fixed bug with image in admin panel, change styles, refactoring, fixing bugs in host page, created better hostNotes, new gameRules, optimization for tests

+++ src/routes/host/PlayerActions.svelte
@@ -12,7 +12,7 @@
 </script>
 
 <div class="actions">
-	{#if $allowToManipulate}
+	{#if $allowToManipulate && onDelete != null}
 		<button class="delete-btn" on:click={() => onDelete(index)}>
 			<Trash2 size={20} style="stroke: #ff4a4a;" />
 		</button>
